z64 -> Zelda64, clipboard

+++ include/Vector.h
@@ -76,10 +76,10 @@ typedef struct {
 } VecSph;
 
 typedef struct {
-	f64 x;
-	f64 y;
-	f64 w;
-	f64 h;
+	f32 x;
+	f32 y;
+	f32 w;
+	f32 h;
 } Rectf32;
 
 typedef struct {
